modularize resultwriter class
