+++ server/resultwriter.py
@@ -1,5 +1,4 @@
 #!/usr/bin/python
-import subprocess
 import pika
 from json import loads
 from threading import Thread
@@ -8,14 +7,12 @@ import logging
 
 class ResultWriter(Thread):
 
-    def __init__(self, rabbit_host, result_exchange, task_exchange,
-                 icingacmd_path="/var/run/icinga2/cmd/icinga2.cmd"):
+    def __init__(self, rabbit_host, result_exchange, task_exchange):
         Thread.__init__(self)
         self.__running = False
         self._config = {"rabbit_host": rabbit_host,
                         "result_exchange": result_exchange,
-                        "task_exchange": task_exchange,
-                        "icingacmd_path": icingacmd_path}
+                        "task_exchange": task_exchange}
 
     def run(self):
         self.__running = True
@@ -56,41 +53,14 @@ class ResultWriter(Thread):
             )
 
     def _message_callback(self, channel, method, properties, body_json):
-        body = loads(body_json)
-        logging.debug("Received message from rabbit: "+str(body))
-        icingacmd_string = self._get_icingacmd_string(body)
-        icingacmd_commandline_string = self._get_icingacmd_commandline_string(icingacmd_string)
-        self._execute_command(icingacmd_commandline_string)
+        result = loads(body_json)
+        logging.debug("Received message from rabbit: "+str(result))
+        processed_result = self._process_result(result)
+        self._write_result(processed_result)
 
-    def _get_icingacmd_string(self, check_result):
-        if check_result["icingacmd_type"] == "PROCESS_SERVICE_CHECK_RESULT":
-            username = check_result["hostname"].split("_")[0]
-            icingacmd_string = "[{0}] PROCESS_SERVICE_CHECK_RESULT;{1};{5}_{2};{3};{4}" \
-                .format(check_result["time"],
-                        check_result["hostname"],
-                        check_result["name"],
-                        check_result["severity_code"],
-                        check_result["output"],
-                        username)
-        else:
-            icingacmd_string = "[{0}] PROCESS_HOST_CHECK_RESULT;{1};{2};{3}" \
-                .format(check_result["time"],
-                        check_result["hostname"],
-                        check_result["severity_code"],
-                        check_result["output"])
-        return icingacmd_string
+    def _process_result(self, check_result):
+        result_string = str(check_result)
+        return result_string
 
-    def _get_icingacmd_commandline_string(self, icingacmd_string):
-        return "echo '"+icingacmd_string+"' >> "+self._config["icingacmd_path"]+";"
-
-    def _execute_command(self, command_string):
-        logging.info("Execute command: "+command_string)
-        try:
-            output = subprocess.check_output(command_string, shell=True)
-            returncode = 0
-        except subprocess.CalledProcessError as result:
-            output = result.output
-            returncode = result.returncode
-        finally:
-            logging.debug("Outut: "+output+"; code: "+str(returncode))
-            return {"output": output, "returncode": returncode}
+    def _write_result(self, result):
+        print(result)
